独自のフックを作成してみよう
