+++ 12_hooks_p2/src/070_customHook/start/Example.jsx
@@ -1,79 +1,35 @@
-import { useEffect, useState, useLayoutEffect } from "react";
+import { useState } from "react";
+import useTimer from "./useTimer";
 
+// POINT カスタムフックを利用した実装
 const Example = () => {
   const [isDisp, setIsDisp] = useState(true);
 
   return (
     <>
-      {isDisp && <Timer/>}
-      <button onClick={() => setIsDisp(prev => !prev)}>{isDisp ? '非表示' : '表示'}</button>
+      {isDisp && <Timer />}
+      <button onClick={() => setIsDisp((prev) => !prev)}>
+        {isDisp ? "非表示" : "表示"}
+      </button>
     </>
-  )
-}
+  );
+};
 
 const Timer = () => {
-  const [time, setTime] = useState(0);
-  const [isRunning, setIsRunning] = useState(false);
-
-  useEffect(() => {
-    // console.log('init');
-    let intervalId = null;
-
-    if(isRunning) {
-      // console.log('timer start');
-
-      intervalId = window.setInterval(() => {
-        // console.log('interval running');
-        setTime(prev => prev + 1);
-      }, 1000);
-    }
-    
-    return () => {
-      window.clearInterval(intervalId)
-      // console.log('end');
-    }
-  }, [isRunning])
-  
-  useEffect(() => {
-    // // console.log('updated');
-    
-    document.title = 'counter:' + time;
-    window.localStorage.setItem('time-key', time);
-
-    return () => {
-      // debugger
-      // // console.log('updated end');
-    }
-  }, [time]);
-
-  useLayoutEffect(() => {
-    const _time = parseInt(window.localStorage.getItem('time-key'));
-    if(!isNaN(_time)) {
-      setTime(_time);
-    }
-  }, [])
-
-  const toggle = () => {
-    setIsRunning(prev => !prev);
-  }
-
-  const reset = () => {
-    setTime(0);
-    setIsRunning(false);
-  }
+  const { time, isRunning, toggle, reset } = useTimer();
 
   return (
     <>
-    <h3>
-      <time>{time}</time>
-      <span>秒経過</span>
-    </h3>
-    <div>
-      <button onClick={toggle}>{isRunning ? '一時停止' : 'スタート'}</button>
-      <button onClick={reset}>リセット</button>
-    </div>
+      <h3>
+        <time>{time}</time>
+        <span>秒経過</span>
+      </h3>
+      <div>
+        <button onClick={toggle}>{isRunning ? "一時停止" : "スタート"}</button>
+        <button onClick={reset}>リセット</button>
+      </div>
     </>
-    );
+  );
 };
 
 export default Example;
