+++ 12_hooks_p2/src/070_customHook/start/useTimer.js
@@ -1,0 +1,65 @@
+import { useState, useEffect, useLayoutEffect } from "react";
+
+//use〇〇と定義されるのが、カスタムフックと呼ばれる
+const useTimer = () => {
+  const Timer = () => {
+    const [time, setTime] = useState(0);
+    const [isRunning, setIsRunning] = useState(false);
+
+    useEffect(() => {
+      // console.log('init');
+      let intervalId = null;
+
+      if (isRunning) {
+        // console.log('timer start');
+
+        intervalId = window.setInterval(() => {
+          // console.log('interval running');
+          setTime((prev) => prev + 1);
+        }, 1000);
+      }
+
+      return () => {
+        window.clearInterval(intervalId);
+        // console.log('end');
+      };
+    }, [isRunning]);
+
+    useEffect(() => {
+      // // console.log('updated');
+
+      document.title = "counter:" + time;
+      window.localStorage.setItem("time-key", time);
+
+      return () => {
+        // debugger
+        // // console.log('updated end');
+      };
+    }, [time]);
+
+    useLayoutEffect(() => {
+      const _time = parseInt(window.localStorage.getItem("time-key"));
+      if (!isNaN(_time)) {
+        setTime(_time);
+      }
+    }, []);
+
+    const toggle = () => {
+      setIsRunning((prev) => !prev);
+    };
+
+    const reset = () => {
+      setTime(0);
+      setIsRunning(false);
+    };
+  };
+
+  return {
+    toggle,
+    reset,
+    time,
+    isRunning,
+  };
+};
+
+export default useTimer;
